+++ src/hooks/useMyList.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { MyListContext } from '../context/MyListContext.jsx';
+
+export const useMyList = () => {
+  return useContext(MyListContext);
+};
